release: fail closed on every storage failure, alpha.21
A database outage used to be read as an answer. A failed policy, feature,
role, session, or second-factor read continued as if the setting were off,
as if the account held no role, or as if the request were a guest, and
report submission redirected as if the report had been filed while the
insert had actually failed.

These paths now answer 503, and the whole application uses 503 for a
storage failure so an operator can tell a dependency outage from a bug.
docs/operations.md documents the behavior, the error_chain_kind log field,
and the recovery steps.

Reporting is fail-closed end to end: the feature flag, the target lookup,
and the write each answer 503, while a genuinely missing target still
answers 404.

Two aggregates decoded SUM(bigint) as i64, which PostgreSQL returns as
NUMERIC, so they failed on every call. The per-account second-factor guess
limit added in alpha.18 therefore never applied, and the sidebar reply
count was always zero. Both now cast in SQL.

The sidebar recent-threads list cost one query per row on every page
render, and the old per-row query could surface soft-deleted threads. It is
now a single JOIN that filters them. Migration 0005 adds a partial index
for the posting cooldown, which had no index to serve it.

The E2E suite asserts 503 for twelve distinct storage failures over the
real router, each confirming the fault landed on the statement it intends,
and writes a credential-free artifact that CI uploads. The startup smoke
test and the backup/upgrade suite resolved the binary through a hard-coded
./target path, so both reported a false failure wherever cargo is
configured with a build.target-dir; they now ask cargo where it builds.

+++ src/store.rs
@@ -1684,8 +1684,12 @@ impl Store {
         user_id: i64,
         since: DateTime<Utc>,
     ) -> anyhow::Result<i64> {
+        // `SUM(bigint)` returns NUMERIC in PostgreSQL, which sqlx cannot decode
+        // into an i64, so the previous query failed on every call and the
+        // caller silently read the error as "no failed attempts". Cast in SQL
+        // instead of decoding NUMERIC so the per-account limit actually works.
         let row: (i64,) = sqlx::query_as(
-            "SELECT COALESCE(SUM(attempts), 0) FROM pending_logins \
+            "SELECT COALESCE(SUM(attempts), 0)::bigint FROM pending_logins \
              WHERE user_id=$1 AND created_at > $2",
         )
         .bind(user_id)
